+++ src/index.ts
@@ -1,7 +1,8 @@
 /*
 The SDK interface for games to interact with Rune.
 */
-import { setupBrowser } from "./setupBrowser"
+import { setupBrowser } from "./internal/setupBrowser"
+import { disableLocalStorage } from "./internal/disableLocalStorage"
 
 interface InitInput {
   startGame: () => void
@@ -192,6 +193,7 @@ export const Rune: RuneExport = {
   },
 }
 
+disableLocalStorage()
 setupBrowser()
 
 // Global namespace properties needed for communicating with Rune
